requirements.txt
 streamlit 
 langchain 
 langchain-groq 
 pandas
 numpy
 python-dotenv
